requirements.txt
streamlit
tensorflow
pillow
numpy
tensorflow-gpu
